refactor(coding-agent): split websearch provider modules
